Split out fixtures tests
